Add custom error response codes

+++ src/api/user/controller/user.controller.ts
@@ -24,6 +24,7 @@ import {
 	ApiOkResponse,
 	ApiTags,
 } from '@nestjs/swagger';
+import { customCodes } from '../../../utils/constants';
 
 @ApiTags('user')
 @Controller('user')
@@ -37,6 +38,16 @@ export class UserController {
 	@ApiForbiddenResponse({ description: 'Forbidden.' })
 	async create(@Body() createUserDto: CreateUserDto) {
 		try {
+			const userFind = await this.userService.findOne(createUserDto?.id);
+			if (userFind) {
+				return {
+					statusCode: HttpStatus.FORBIDDEN,
+					customCode: 'r0003',
+					customMessage: customCodes?.r0003?.description,
+					message: 'User already exist',
+				};
+			}
+
 			const user = await this.userService.create(createUserDto);
 			return {
 				statusCode: HttpStatus.CREATED,
@@ -47,6 +58,8 @@ export class UserController {
 			throw new HttpException(
 				{
 					statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
+					customCode: 'r0013',
+					customMessage: customCodes.r0013?.description,
 					message: `Error creating user: ${error.message}`,
 				},
 				HttpStatus.INTERNAL_SERVER_ERROR
@@ -63,7 +76,12 @@ export class UserController {
 		try {
 			const user = await this.userService.findOne(id);
 			if (!user) {
-				throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+				return {
+					statusCode: HttpStatus.NOT_FOUND,
+					customCode: 'r0002',
+					customMessage: customCodes.r0002?.description,
+					message: 'User not found',
+				};
 			}
 			return {
 				statusCode: HttpStatus.OK,
@@ -78,6 +96,8 @@ export class UserController {
 				{
 					statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
 					message: `Error retrieving user: ${error.message}`,
+					customCode: 'r0016',
+					customMessage: customCodes.r0016?.description,
 				},
 				HttpStatus.INTERNAL_SERVER_ERROR
 			);
@@ -91,6 +111,15 @@ export class UserController {
 	@ApiForbiddenResponse({ description: 'Forbidden.' })
 	async update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
 		try {
+			const userFind = await this.userService.findOne(id);
+			if (!userFind) {
+				return {
+					statusCode: HttpStatus.NOT_FOUND,
+					customCode: 'r0002',
+					customMessage: customCodes.r0002?.description,
+					message: 'User not found',
+				};
+			}
 			const user = await this.userService.update(id, updateUserDto);
 			return {
 				statusCode: HttpStatus.OK,
@@ -102,6 +131,8 @@ export class UserController {
 				{
 					statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
 					message: `Error updating user: ${error.message}`,
+					customCode: 'r0016',
+					customMessage: customCodes.r0016?.description,
 				},
 				HttpStatus.INTERNAL_SERVER_ERROR
 			);
@@ -115,6 +146,15 @@ export class UserController {
 	@ApiForbiddenResponse({ description: 'Forbidden.' })
 	async remove(@Param('id') id: string) {
 		try {
+			const userFind = await this.userService.findOne(id);
+			if (!userFind) {
+				return {
+					statusCode: HttpStatus.NOT_FOUND,
+					customCode: 'r0002',
+					customMessage: customCodes.r0002?.description,
+					message: 'User not found',
+				};
+			}
 			await this.userService.remove(id);
 			return {
 				statusCode: HttpStatus.OK,
@@ -125,6 +165,8 @@ export class UserController {
 				throw new HttpException(
 					{
 						statusCode: HttpStatus.NOT_FOUND,
+						customCode: 'r0002',
+						customMessage: customCodes.r0002?.description,
 						message: error.message,
 					},
 					HttpStatus.NOT_FOUND
@@ -134,6 +176,8 @@ export class UserController {
 					{
 						statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
 						message: `Error deleting user: ${error.message}`,
+						customCode: 'r0016',
+						customMessage: customCodes.r0016?.description,
 					},
 					HttpStatus.INTERNAL_SERVER_ERROR
 				);
@@ -148,6 +192,15 @@ export class UserController {
 	@ApiForbiddenResponse({ description: 'Forbidden.' })
 	async signin(@Body() signinDto: SignInDto) {
 		try {
+			const userFind = await this.userService.findOneByEmail(signinDto?.email);
+			if (!userFind) {
+				return {
+					statusCode: HttpStatus.NOT_FOUND,
+					customCode: 'r0002',
+					customMessage: customCodes.r0002?.description,
+					message: 'User not found',
+				};
+			}
 			const result = await this.userService.signin(signinDto);
 			return {
 				statusCode: HttpStatus.OK,
@@ -158,6 +211,8 @@ export class UserController {
 			throw new HttpException(
 				{
 					statusCode: HttpStatus.UNAUTHORIZED,
+					customCode: 'r0001',
+					customMessage: customCodes.r0001?.description,
 					message: error.message,
 				},
 				HttpStatus.UNAUTHORIZED
@@ -175,6 +230,17 @@ export class UserController {
 		@Body() authChangePasswordUserDto: AuthChangePasswordUserDto
 	) {
 		try {
+			const userFind = await this.userService.findOneByEmail(
+				authChangePasswordUserDto?.email
+			);
+			if (!userFind) {
+				return {
+					statusCode: HttpStatus.NOT_FOUND,
+					customCode: 'r0002',
+					customMessage: customCodes.r0002?.description,
+					message: 'User not found',
+				};
+			}
 			const message = await this.userService.changeUserPassword(
 				authChangePasswordUserDto
 			);
@@ -187,6 +253,8 @@ export class UserController {
 				{
 					statusCode: HttpStatus.BAD_REQUEST,
 					message: error,
+					customCode: 'r0016',
+					customMessage: customCodes.r0016?.description,
 				},
 				HttpStatus.BAD_REQUEST
 			);
@@ -203,6 +271,17 @@ export class UserController {
 		@Body() authForgotPasswordUserDto: AuthForgotPasswordUserDto
 	) {
 		try {
+			const userFind = await this.userService.findOneByEmail(
+				authForgotPasswordUserDto?.email
+			);
+			if (!userFind) {
+				return {
+					statusCode: HttpStatus.NOT_FOUND,
+					customCode: 'r0002',
+					customMessage: customCodes.r0002?.description,
+					message: 'User not found',
+				};
+			}
 			const message = await this.userService.forgotUserPassword(
 				authForgotPasswordUserDto
 			);
@@ -215,6 +294,8 @@ export class UserController {
 				{
 					statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
 					message: error,
+					customCode: 'r0016',
+					customMessage: customCodes.r0016?.description,
 				},
 				HttpStatus.INTERNAL_SERVER_ERROR
 			);
@@ -231,6 +312,17 @@ export class UserController {
 		@Body() authConfirmPasswordUserDto: AuthConfirmPasswordUserDto
 	) {
 		try {
+			const userFind = await this.userService.findOneByEmail(
+				authConfirmPasswordUserDto?.email
+			);
+			if (!userFind) {
+				return {
+					statusCode: HttpStatus.NOT_FOUND,
+					customCode: 'r0002',
+					customMessage: customCodes.r0002?.description,
+					message: 'User not found',
+				};
+			}
 			const message = await this.userService.confirmUserPassword(
 				authConfirmPasswordUserDto
 			);
@@ -243,6 +335,8 @@ export class UserController {
 				{
 					statusCode: HttpStatus.INTERNAL_SERVER_ERROR,
 					message: error,
+					customCode: 'r0016',
+					customMessage: customCodes.r0016?.description,
 				},
 				HttpStatus.INTERNAL_SERVER_ERROR
 			);
